user_profile -> userprofile refactor
